phase2: extract bridge_observability.rs and bridge_sse_helpers.rs from bridge_inprocess
- bridge_observability.rs (174 lines): context trace signals, quality assessment persistence
- bridge_sse_helpers.rs (293 lines, 8 tests): render_sse, apply_forward_llm_sse_event,
  extend_forward_from_validated_sse_block, flush_tail_buf_into_llm_forward
- Remove duplicate tests from bridge_inprocess (render_sse_map, reasoning_done)
- Clean dead imports (SessionQualityAssessmentRequest, ContextTrace*, Database*EventService)
- bridge_inprocess.rs: 5247 → 4943 lines (-304)
- Total extracted from original 5920-line bridge: ~1,263 lines across 4 modules
